+++ silverShelter/src/main/java/com/silver/shelter/admin/model/service/AdminServiceImpl.java
@@ -1,0 +1,71 @@
+package com.silver.shelter.admin.model.service;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.springframework.stereotype.Service;
+
+import com.silver.shelter.admin.model.mapper.AdminMapper;
+import com.silver.shelter.member.model.dto.Member;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class AdminServiceImpl implements AdminService{
+
+	private final AdminMapper mapper;
+	
+	
+	// 회원 목록 조회
+	@Override
+	public Map<String, Object> memberALlSelect() {
+		
+		List<Member> memberList = mapper.memberAllSelect();
+		
+		Map<String, Object> map = new HashMap<>();
+		
+		map.put("memberList", memberList);
+		
+		return map;
+		
+	}
+	
+	
+
+	// 검색한 회원 목록 조회
+	@Override
+	public Map<String, Object> memberSearchSelect(Map<String, Object> paramMap) {
+		
+//		int listCount = mapper.getSearchCount(paramMap);
+		
+		List<Member> boardList = mapper.memberSearchSelect(paramMap);
+		
+		Map<String, Object> map = new HashMap<>();
+		
+		map.put("boardList", boardList);
+		
+		return map;
+	}
+
+
+	// 회원 상세 조회
+	@Override
+	public Member adminDetailSelect(Member member) {
+		
+		Member memberList = mapper.adminDetailSelect(member);
+		
+		return memberList;
+	}
+	
+	
+	
+	
+	
+	
+	
+	
+	
+
+}
