+++ silverShelter/src/main/java/com/silver/shelter/board/model/dto/Pagination.java
@@ -1,0 +1,12 @@
+package com.silver.shelter.board.model.dto;
+
+public class Pagination {
+
+	
+	private int currentPage;     // 현제 페이지 번호
+	private int listCount;       // 전체 게시글 수
+	
+	private int limit = 10;      // 한 페이
+	
+	
+}
