+++ silverShelter/src/main/java/com/silver/shelter/board/model/service/BoardService.java
@@ -1,0 +1,5 @@
+package com.silver.shelter.board.model.service;
+
+public interface BoardService {
+
+}
