+++ silverShelter/src/main/java/com/silver/shelter/common/config/SecurityConfig.java
@@ -1,0 +1,15 @@
+package com.silver.shelter.common.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+@Configuration
+public class SecurityConfig {
+
+	@Bean
+	public BCryptPasswordEncoder bCryptPasswordEncoder() {
+		
+		return new BCryptPasswordEncoder();
+	}
+}
