+++ silverShelter/src/main/java/com/silver/shelter/communication/controller/CommunicationController.java
@@ -1,0 +1,16 @@
+package com.silver.shelter.communication.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("communication")
+public class CommunicationController {
+
+	@GetMapping("communication")
+	public String communication() {
+		
+		return "communication/communication";
+	}
+}
