+++ silverShelter/src/main/java/com/silver/shelter/member/model/mapper/MemberMapper.java
@@ -1,0 +1,8 @@
+package com.silver.shelter.member.model.mapper;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface MemberMapper {
+
+}
