+++ silverShelter/src/main/java/com/silver/shelter/member/model/service/MemberService.java
@@ -1,0 +1,5 @@
+package com.silver.shelter.member.model.service;
+
+public interface MemberService {
+
+}
