+++ silverShelter/src/main/java/com/silver/shelter/member/model/service/MemberServiceImpl.java
@@ -1,0 +1,16 @@
+package com.silver.shelter.member.model.service;
+
+import org.springframework.stereotype.Service;
+
+import com.silver.shelter.member.model.mapper.MemberMapper;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Service
+public class MemberServiceImpl implements MemberService{
+
+//	private final MemberMapper mapper;
+	
+	
+}
