+++ silverShelter/src/main/java/com/silver/shelter/signUp/controller/SignUpController.java
@@ -1,0 +1,17 @@
+package com.silver.shelter.signUp.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("signUp")
+public class SignUpController {
+
+	@GetMapping("documentSubmission")
+	public String documentSubmission() {
+		
+		return "signUp/documentSubmission";
+	}
+	
+}
