--- silverShelter/src/main/java/silverShelter/board/model/service/BoardService.java
@@ -1,5 +1,0 @@
-package silverShelter.board.model.service;
-
-public interface BoardService {
-
-}
