--- silverShelter/src/main/java/silverShelter/common/config/EmailConfig.java
@@ -1,45 +1,0 @@
-package silverShelter.common.config;
-
-import java.util.Properties;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
-
-@PropertySource("classpath:/config.properties")
-@Configuration
-public class EmailConfig {
-
-	@Value("${spring.mail.username}")
-	private String userName;
-	
-	@Value("${spring.mail.password}")
-	private String password;	
-	
-	@Bean
-	public JavaMailSender javaMailSender() {
-		
-		JavaMailSenderImpl mailSender = new JavaMailSenderImpl();
-		
-		Properties prop = new Properties();
-		prop.setProperty("mail.transport.protocol", "smtp");
-		prop.setProperty("mail.smtp.auth", "true");
-		prop.setProperty("mail.smtp.starttls.enable", "true");
-		prop.setProperty("mail.debug", "true");
-		prop.setProperty("mail.smtp.ssl.trust","smtp.gmail.com");
-		prop.setProperty("mail.smtp.ssl.protocols","TLSv1.2");
-		
-		
-		mailSender.setUsername(userName);
-		mailSender.setPassword(password);
-		mailSender.setHost("smtp.gmail.com");
-		mailSender.setPort(587);
-		mailSender.setDefaultEncoding("UTF-8");
-		mailSender.setJavaMailProperties(prop);
-		
-		return mailSender;
-	}
-}
