--- silverShelter/src/main/java/silverShelter/common/config/SecurityConfig.java
@@ -1,15 +1,0 @@
-package silverShelter.common.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@Configuration
-public class SecurityConfig {
-
-	@Bean
-	public BCryptPasswordEncoder bCryptPasswordEncoder() {
-		
-		return new BCryptPasswordEncoder();
-	}
-}
