--- silverShelter/src/main/java/silverShelter/main/controller/MainController.java
@@ -1,16 +1,0 @@
-package silverShelter.main.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class MainController {
-	
-	@RequestMapping("/")
-	public String mainPage() {
-		return "common/main";
-	}
-	
-
-}
