--- silverShelter/src/main/java/silverShelter/member/model/mapper/MemberMapper.java
@@ -1,8 +1,0 @@
-package silverShelter.member.model.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface MemberMapper {
-
-}
