--- silverShelter/src/main/java/silverShelter/member/model/service/MemberService.java
@@ -1,5 +1,0 @@
-package silverShelter.member.model.service;
-
-public interface MemberService {
-
-}
