--- silverShelter/src/main/java/silverShelter/member/model/service/MemberServiceImpl.java
@@ -1,15 +1,0 @@
-package silverShelter.member.model.service;
-
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-import silverShelter.member.model.mapper.MemberMapper;
-
-@RequiredArgsConstructor
-@Service
-public class MemberServiceImpl implements MemberService{
-
-//	private final MemberMapper mapper;
-	
-	
-}
