+++ silverShelter/src/main/resources/config.properties
@@ -70,3 +70,11 @@ my.board.resource-location=file:///C:/norazoServer/board/
 my.board.web-path=/images/board/
 # 게시글 이미지를 서버에 저장할 때 사용할 경로 ( transferTo() ) 
 my.board.folder-path=C:/norazoServer/board/
+
+# -------------------------------------------------------------
+
+# 공공데이터 포털 ServiceKey encode
+my.public.data.service.key.encode=QH%2FFLIzW5Mvv5mhKZOBq5jo9y2RhH8A%2FrKTdDcntOdAmZagTpf0y4pYiiPw%2FNumb6Qm8RATvL9bW8GFD6QN1dQ%3D%3D
+
+# 공공데이터 포털 ServiceKey decode(js 에서 URLSearchParams 사용 시 decode 키 사용해야함)
+my.public.data.service.key.decode=QH/FLIzW5Mvv5mhKZOBq5jo9y2RhH8A/rKTdDcntOdAmZagTpf0y4pYiiPw/Numb6Qm8RATvL9bW8GFD6QN1dQ==
