+++ silverShelter/src/main/resources/static/js/common/main.js
@@ -1,21 +1,3 @@
-   /* 네비창 script */ 
-/*
-document.addEventListener('DOMContentLoaded', function() {
-    
-    document.querySelector('.hidden-bar').addEventListener('click', function() {
-
-        	document.getElementById('sideNav').classList.toggle('active');
-    });
-
-
-        document.querySelector('.closebtn').addEventListener('click', function() {
-
-        document.getElementById('sideNav').classList.remove('active');
-
-    	});
-});
-*/
-
 document.addEventListener('DOMContentLoaded', function() {
 
 	const indicators = document.querySelectorAll('.indicator');
@@ -93,3 +75,124 @@ document.addEventListener('DOMContentLoaded', function() {
 }); /* 캐러샐2 end */
     
  
+// 날씨 API 
+
+function getCurrentDate(){
+    const today = new Date();
+    const year = today.getFullYear();
+    const month = ('0' + (today.getMonth() + 1)).slice(-2);
+    const day = ('0'+ (today.getDate())).slice(-2);
+    
+    return `${year}${month}${day}`;
+}
+
+console.log(getCurrentDate());
+
+
+async function getServiceKey() {
+    
+    try{
+
+        const response = await fetch("/getServiceKey");
+        
+        return response.text();
+        
+
+    } catch(err){
+        console.log("getServiceKey의 에러" + err)
+    }
+}
+
+// 공공데이터 날씨 API 정보를 얻어올 함수 
+async function fetchData(){
+    
+    // 날씨 이미지 
+    const weatherIcon = document.getElementById("weatherIcon")
+
+    const skyInformation = document.getElementById("skyInformation")
+    
+    const temperature = document.getElementById("temperature")
+    
+    const precipitation = document.getElementById("precipitation")
+
+    const CurrentDate = getCurrentDate();
+
+    const servicekey = await getServiceKey(); // 비동기 요청 1번째의 응답이 올대까지 기다림 
+
+    console.log("serviceKey : ", servicekey);
+
+    const url = 'http://apis.data.go.kr/1360000/VilageFcstInfoService_2.0/getVilageFcst';
+
+    //URLSearchParams : URL의 쿼리문자열을 쉽게 다룰 수 있게 해주는 내장 객체 
+    // 단, 사용시 decode 서비스키 사용 -> URLSearchParams이 데이터를 인코딩하기 때문 
+    const queryParams = new URLSearchParams({
+        
+        serviceKey : servicekey,
+        pageNo : 1, 
+        numOfRows : 10,
+        dataType : 'JSON',
+        base_date : CurrentDate,
+        base_time : '0500',
+        nx : 60,
+        ny: 127 
+
+    });
+
+    console.log(`${url}?${queryParams}`);
+
+    // fetch 요청 
+
+    // 비동기 요청 2번째 
+    try{
+        const response = await fetch(`${url}?${queryParams}`);
+        const result = await response.json();
+
+        console.log(result);
+
+
+        const obj = result.response.body.items.item.reduce((acc, data) => {
+            acc[data.category] = data.fcstValue;
+            return acc;
+        }, {});
+   
+        // console.log(obj);
+        const sky = {
+            "1" : "맑음",
+            "3" : "구름많음",
+            "4" : "흐림"
+        }
+        
+        if( obj.PTY == 0 ) { // 강수 없음
+            switch(obj.SKY) { // SKY 상태에 따라 아이콘 지정
+                case "1":
+                    weatherIcon.src = '/images/weather/sunny.png'; // 맑음 이미지 파일 경로
+                    break;
+                case "3":
+                    weatherIcon.src = '/images/weather/cloudy.png'; // 구름 많음 이미지 파일 경로
+                    break;
+                case "4":
+                    weatherIcon.src = '/images/weather/overcast.png'; // 흐림 이미지 파일 경로
+                    break;
+   
+            }
+        } else if(obj.PTY == 3) { // 눈 올 때
+            weatherIcon.src = 'images/weather/snow.png';
+        } else { // 그외 비올때
+            weatherIcon.src = 'images/weather/rain.png';
+        }
+        // 하늘 정보
+        skyInformation.innerText = sky[obj.SKY];
+        // 기온
+        temperature.innerText = `${obj.TMP}℃`;
+       
+        // 강수확률
+        precipitation.innerText = `강수 확률 : ${obj.POP}%`;
+
+
+    } catch(err){
+
+        console.log(err);
+    }
+}
+
+fetchData();
