+++ silverShelter/src/main/resources/templates/common/common.html
@@ -16,4 +16,3 @@
 <link rel="stylesheet" href="/css/common/headerNav.css">
 
 
-
