+++ silverShelter/src/main/resources/templates/common/mainNav.html
@@ -5,10 +5,10 @@
     
     <div class="nav-line"></div>
     <div>
-        <a href="/member/Introduction">입주정보</a>
+        <a href="/introduction/introduction">입주정보</a>
     </div>
     <div>
-        <a href="#">시설소개</a>
+        <a href="/communication/communication">시설소개</a>
     </div>
     <div>
         <a href="#">의료센터</a>
