+++ silverShelter/src/main/resources/templates/introduction/introductionCommon.html
@@ -5,6 +5,6 @@
         
         <!-- 시설 관련 div -->
         <div class="facility-links">
-            <a href="/introduction/introduction">단지소개</a>
-            <a href="/introduction/moveInGuide">입주 안내</a>
+            <a href="/introduction/introduction" id="introductionABtn">단지소개</a>
+            <a href="/introduction/moveInGuide" id="moveInGuideABtn">입주 안내</a>
         </div>
